perf(plugin): move all lazy.nvim related code to `lazyvim.util.plugin`
