+++ lua/lazyvim/plugins/init.lua
@@ -1,3 +1,13 @@
+if vim.fn.has("nvim-0.9.0") == 0 then
+  vim.api.nvim_echo({
+    { "LazyVim requires Neovim >= 0.9.0\n", "ErrorMsg" },
+    { "Press any key to exit", "MoreMsg" },
+  }, true, {})
+  vim.fn.getchar()
+  vim.cmd([[quit]])
+  return {}
+end
+
 require("lazyvim.config").init()
 
 return {
